requirements.txt
flask==3.0.0
flask-cors==4.0.0
selenium==4.15.2
gspread==5.12.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pandas==2.1.4
webdriver-manager==4.0.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
gunicorn==21.2.0

